requirements.txt
asgiref==3.6.0
attrs==22.2.0
boto3==1.26.95
botocore==1.29.95
Django==4.1.7
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
psycopg2==2.9.5
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
